backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai>=1.5.0
google-cloud-texttospeech==2.22.0
pydantic==2.10.4
